Do not restart buckets automatically
The dynamic supervisor used to start bucket processes
normally would restart a bucket if it dies.  For this
application this is not ideal, as the regisrty would
not know of the newly started bucket process.  This
means that we'd have a bucket which would be inaccessible.

To fix this, we configure the bucket agent's restart type to
`temporary`, which cases the suprvisor to not restart the
agent if it dies.