Supervior Strategy fix
When the registry dies, we need to restart the
bucket supervisor, as we just lost all refences to
buckets.

What we want is that the bucket supervisor is restarted if
the registry crashes.  We can't use `:rest_for_all` as the
ordering is (1) bucket supervisor and (2) registry.

The only option left is `:one_for_all`, which causes the supervisor
to restart all children.